requirements.txt
torch>=2.0.0
xfuser
pillow
numpy
einops 